requirements.txt
Django==4.1.5
python-dotenv==0.21.0
djangorestframework==3.14.0
django-cors-headers==3.13.0